liran cr
