--- mlrun/api/launcher/server.py
@@ -1,43 +1,0 @@
-# Copyright 2023 MLRun Authors
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-from mlrun.launcher.abstract import AbstractLauncher
-
-
-class ServerSideLauncher(AbstractLauncher):
-    def __init__(self, **kwargs):
-        super().__init__(**kwargs)
-
-    @staticmethod
-    def run(runtime):
-        pass
-
-    @staticmethod
-    def verify_base_image(runtime):
-        pass
-
-    @staticmethod
-    def save(runtime):
-        pass
-
-    @staticmethod
-    def _enrich_and_validate(runtime):
-        """
-        Enrich the function with:
-            1. Default values
-            2. mlrun config values
-            3. Project context values
-            4. Run specific parameters
-         and validate
-        """
-        pass
