+++ mlrun/launcher/remote.py
@@ -11,3 +11,32 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+from mlrun.launcher.base import _BaseLauncher
+
+
+class RemoteLauncher(_BaseLauncher):
+    @staticmethod
+    def verify_base_image(runtime):
+        pass
+
+    @staticmethod
+    def save(runtime):
+        pass
+
+    @staticmethod
+    def run(runtime):
+        pass
+
+    @staticmethod
+    def _enrich_runtime(runtime):
+        pass
+
+    @staticmethod
+    def _validate_runtime(runtime):
+        pass
+
+    def _submit_job(self, runtime):
+        pass
+
+    def _deploy(self, runtime):
+        pass
